17 4 done

+++ src/lib/year_2017/day_04/part_01.rs
@@ -1,7 +1,7 @@
 use std::collections::HashSet;
 
-pub fn part_01(input: &[&'static str]) -> usize {
-    input.into_iter().fold(0usize, |acc, line| {
+pub fn part_01(input: &[&'static str]) -> u16 {
+    input.into_iter().fold(0u16, |acc, line| {
         let mut seen = HashSet::<&[u8]>::new();
         for word in line.split_whitespace().map(|w| w.as_bytes()) {
             if seen.contains(word) {
